Update lab4.2.c
Change int to float

+++ Lab 4/2/lab4.2.c
@@ -1,9 +1,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 int main(){
-	int avg=0, sum=0, n=0, i, h=0;
+	float avg=0, sum=0, n=0, i, h=0;
 	printf("Enter the size of the array: ");
-	scanf("%d", &n);
+	scanf("%f", &n);
 	int a[n];
 	for(i=0; i<n; i++){
 		a[i]=rand()%10;
@@ -13,7 +13,7 @@ int main(){
 	avg=sum/n;
 	for (i=0; i<n; i++)
 		if (a[i]>avg) h++;
-	printf("The number of elements is more than average: %d\n", h);
+	printf("The number of elements is more than average: %f\n", h);
 	system("pause");
 	return 0;
 }
